Split study-view-screenshot.spec.ts into 3 files for shard balance
The original file held 11 tests in 5 describe blocks and consistently
landed in the slowest shard (172-226s of test exec across recent
runs). Playwright's --shard distributes by file with test-count
balancing, so splitting one heavy file into 3 lets the sharder
spread its work across more parallel runners.

Split respects the `.serial` describes (their tests share chart
state via beforeAll/afterAll and must stay in one file):

  study-view-screenshot-custom-data.spec.ts
    - study view generic assay categorical/binary features (1 skipped)
    - Test the Custom data tab (2 tests, no screenshots)

  study-view-screenshot-x-vs-y.spec.ts
    - study view x vs y charts (3 tests, .serial, 3 screenshots)
    - study view editable breadcrumbs (1 test, .serial, 1 screenshot)

  study-view-screenshot-cancer-gene-filter.spec.ts
    - cancer gene filter (3 tests, .serial, 2 screenshots)

The 6 existing snapshot baselines are git-mv'd to their new spec
subdirectories so toHaveScreenshot lookups continue to match.
